requirements.txt
pyrogram==2.0.106
python-decouple==3.8
alembic==1.13.3
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
asyncpg==0.29.0
